Added support for constraining authentication to a GitHub organization.
This patch will allow users to add an optional GITHUB_ORGANIZATION
setting to their settings.py file which will constrain authentication
to the defined GitHub organization.

If no GitHub organization is defined, authentication will not be
constrained and the backend will work as before.

Also updated the documentation to explain option usage of the
GITHUB_ORGANIZATION setting.